# Conflicts:
#	VisualPinball.Engine.Test/VPT/Primitive/PrimitiveMeshTests.cs

+++ VisualPinball.Engine/Common/Constants.cs
@@ -1,3 +1,19 @@
+// Visual Pinball Engine
+// Copyright (C) 2020 freezy and VPE Team
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <https://www.gnu.org/licenses/>.
+
 // ReSharper disable CommentTypo
 
 namespace VisualPinball.Engine.Common
@@ -37,12 +53,12 @@ namespace VisualPinball.Engine.Common
 		/// <summary>
 		/// step time in seconds
 		/// </summary>
-		public const float PhysicsStepTimeS = PhysicsStepTime * 1e-6f;         // PHYSICS_STEPTIME_S
+		public const double PhysicsStepTimeS = PhysicsStepTime * 1e-6;         // PHYSICS_STEPTIME_S
 
 		/// <summary>
 		/// default physics rate: 1000Hz
 		/// </summary>
-		public const float DefaultStepTime = 10000f;                           // DEFAULT_STEPTIME
+		public const double DefaultStepTime = 10000;                           // DEFAULT_STEPTIME
 
 		/// <summary>
 		/// default physics rate: 1000Hz
@@ -50,7 +66,7 @@ namespace VisualPinball.Engine.Common
 		public const float DefaultStepTimeS = 0.01f;                           // DEFAULT_STEPTIME_S
 
 
-		public const float PhysFactor = PhysicsStepTimeS / DefaultStepTimeS;   // PHYS_FACTOR
+		public const double PhysFactor = PhysicsStepTimeS / DefaultStepTimeS;  // PHYS_FACTOR
 
 
 		public const float LowNormVel = 0.0001f;                               // C_LOWNORMVEL
